Fixed import master (#17)
* Refactored type definition of sockets to commons

* Fixed errors with socket refactor of types

* Fixed error with importing
