+++ packages/suibian-server/src/sockets/index.ts
@@ -1,13 +1,11 @@
 import express from "express";
 import socketio from "socket.io";
+import { joinRoom, broadcastRoom, createRoom } from "./room";
 import {
-    joinRoom,
+    suibianSocketServer,
     joinRoomPayload,
-    roomMessagePayload,
-    broadcastRoom,
-    createRoom
-} from "./room";
-import { suibianSocketServer } from "@suibian/commons";
+    roomMessagePayload
+} from "@suibian/commons";
 const http = require("http");
 
 export default {
